Tests/LibWeb/Text/expected/css/explicit-inherit-non-inherited-properties.txt

position: absolute
float: none
clear: both
overflow-x: hidden
text-decoration-line: underline
grid-template-columns: [a] 100px [b] 200px
grid-template-rows: [c] 50px
